facewise/requirements.txt
Flask==1.1.4
google-cloud-datastore==2.1.2
protobuf==3.19.4
requests==2.25.1
spacy==3.0
tweepy==4.4.0
markupsafe==2.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl